# lipcoding-docs/openapi-generated-formatted.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Mentor-Mentee Matching API",
        "description": "API for matching mentors and mentees in a mentoring platform",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/me": {
            "get": {
                "tags": [
                    "User Profile"
                ],
                "summary": "Get Current User Info",
                "operationId": "get_current_user_info_api_me_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/api/profile": {
            "put": {
                "tags": [
                    "User Profile"
                ],
                "summary": "Update Profile",
                "operationId": "update_profile_api_profile_put",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "anyOf": [
                                    {
                                        "$ref": "#/components/schemas/UpdateMentorProfileRequest"
                                    },
                                    {
                                        "$ref": "#/components/schemas/UpdateMenteeProfileRequest"
                                    }
                                ],
                                "title": "Profile Data"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/api/images/{role}/{user_id}": {
            "get": {
                "tags": [
                    "User Profile"
                ],
                "summary": "Get Profile Image",
                "operationId": "get_profile_image_api_images__role___user_id__get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "role",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Role"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "User Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "UpdateMenteeProfileRequest": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "role": {
                        "type": "string",
                        "const": "mentee",
                        "title": "Role"
                    },
                    "bio": {
                        "type": "string",
                        "title": "Bio"
                    },
                    "image": {
                        "type": "string",
                        "title": "Image"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "name",
                    "role",
                    "bio",
                    "image"
                ],
                "title": "UpdateMenteeProfileRequest"
            },
            "UpdateMentorProfileRequest": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "role": {
                        "type": "string",
                        "const": "mentor",
                        "title": "Role"
                    },
                    "bio": {
                        "type": "string",
                        "title": "Bio"
                    },
                    "image": {
                        "type": "string",
                        "title": "Image"
                    },
                    "skills": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Skills"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "name",
                    "role",
                    "bio",
                    "image",
                    "skills"
                ],
                "title": "UpdateMentorProfileRequest"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}